requirements.txt
-i https://pypi.org/simple
jinja2==2.10
markupsafe==1.0
numpy==1.14.4
pyaudio==0.2.11
python-decouple==3.1
